Develop

+++ TowerOfWitch/Models/Player.cs
@@ -16,6 +16,6 @@ namespace TowerOfWitch.Models
         public int WinGame { get; set; }
         public uint Coins { get; set; }
         public bool InGame { get; set; }
-        public int SymbolCode { get; set; }
+        public byte SymbolCode { get; set; }
     }
 }
